+++ src/main/java/com/example/shoppingmall_comp/domain/members/controller/MemberController.java
@@ -4,7 +4,7 @@ import com.example.shoppingmall_comp.domain.members.dto.MemberSignInRequest;
 import com.example.shoppingmall_comp.domain.members.dto.MemberSignInResponse;
 import com.example.shoppingmall_comp.domain.members.dto.MemberSignUpRequest;
 import com.example.shoppingmall_comp.domain.members.dto.MemberSignUpResponse;
-import com.example.shoppingmall_comp.domain.members.service.serviceImpl.MemberServiceImpl;
+import com.example.shoppingmall_comp.domain.members.service.Impl.MemberServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
