+++ src/main/java/com/example/shoppingmall_comp/domain/members/controller/TokenController.java
@@ -2,7 +2,7 @@ package com.example.shoppingmall_comp.domain.members.controller;
 
 import com.example.shoppingmall_comp.domain.members.dto.CreateAccessTokenReponse;
 import com.example.shoppingmall_comp.domain.members.dto.CreateAccessTokenRequest;
-import com.example.shoppingmall_comp.domain.members.service.serviceImpl.TokenServiceImpl;
+import com.example.shoppingmall_comp.domain.members.service.Impl.TokenServiceImpl;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
